+++ varinfo/generate_umm_var.py
@@ -17,7 +17,7 @@ from cmr import CMR_UAT
 
 from varinfo import VarInfoFromNetCDF4
 from varinfo.cmr_search import (CmrEnvType, download_granule, get_granule_link,
-                                get_granules)
+                                get_granules, get_edl_token_header)
 from varinfo.umm_var import get_all_umm_var, publish_all_umm_var
 
 
@@ -45,8 +45,11 @@ def generate_collection_umm_var(collection_concept_id: str,
         Note - if attempting to publish to CMR, a LaunchPad token must be used.
 
     """
+    # Get an EDL token and its header given a LaunchPad token
+    auth_header_edl_token = get_edl_token_header(auth_header, cmr_env)
+
     granule_response = get_granules(collection_concept_id, cmr_env=cmr_env,
-                                    auth_header=auth_header)
+                                    auth_header=auth_header_edl_token)
 
     # Get the data download URL for the most recent granule (NetCDF-4 file)
     granule_link = get_granule_link(granule_response)
@@ -54,7 +57,7 @@ def generate_collection_umm_var(collection_concept_id: str,
     with TemporaryDirectory() as temp_dir:
         # Download file to runtime environment
         local_granule = download_granule(granule_link,
-                                         auth_header,
+                                         auth_header_edl_token,
                                          out_directory=temp_dir)
 
         # Parse the granule with VarInfo to map all variables and relations: